requirements.txt
Babel==2.4.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.3
django-auditlog==0.4.3
django-cors-headers==2.1.0
django-csv-export==0.0.3
django-filter==1.0.4
django-import-export==1.0.0
django-jsonfield==1.0.1
django-model-to-dict==0.2.1
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
djangorestframework-bulk==0.2.1
djangorestframework-csv==2.1.0
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
Markdown==2.6.8
odfpy==1.3.6
openpyxl==2.5.1
phonenumberslite==8.7.1
psycopg2==2.7.3
pycparser==2.18
PyJWT==1.5.2
PySocks==1.6.7
python-decouple==3.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
sms==0.4
tablib==0.12.1
twilio==6.6.1
unicodecsv==0.14.1
urllib3==1.22
uuid==1.30
xlrd==1.1.0
xlwt==1.3.0
